requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.9.0
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
openai==0.28.0
pydantic==1.10.13
typing-extensions==4.5.0
